+++ src/Server/Persistence/Database/Sqlite/Queries/SqliteRetrieveAllTemplatesQuery.cs
@@ -37,7 +37,8 @@ public class SqliteRetrieveAllTemplatesQuery : IRetrieveAllTemplatesQuery
                 stackable, quantity, itemUse, npcFlags, useRange,
                 healthValue, healthMaxValue, healthChangeRate, healthChangeInterval,
                 staminaValue, staminaMaxValue, staminaChangeRate, staminaChangeInterval,
-                manaValue, manaMaxValue, manaChangeRate, manaChangeInterval
+                manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
+                statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck
             FROM EntityWithComponents 
             LEFT JOIN EntityKeyValue kv ON kv.entity_id = id
             WHERE id >= @FirstTemplateId AND id <= @LastTemplateId";