+++ src/Server/ServerCore/Entities/ServerEntityBuilder.cs
@@ -58,12 +58,13 @@ public sealed class ServerEntityBuilder : AbstractEntityBuilder
         HealthComponentCollection healths,
         StaminaComponentCollection staminas,
         ManaComponentCollection manas,
+        StatsComponentCollection stats,
         EntityTable entityTable)
         : base(entityId, load, entityManager, kinematics, blockTiles,
             aboveBlocks, playerCharacterTags, names, parents,
             drawables, animatedSprites, orientations, admins, blockPositions, castBlockShadows, pointLightSources,
             physics, boundingBoxes, castShadows, entityTypes, serverOnly, stackables, quantities, itemUses, npcFlags,
-            useRanges, healths, staminas, manas, entityTable)
+            useRanges, healths, staminas, manas, stats, entityTable)
     {
         this.accounts = accounts;
     }